поправил ошибки WP_DEBUG

+++ lib/FS_Api_Class.php
@@ -28,10 +28,10 @@ class FS_Api_Class {
 		global $wpdb;
 		$config = new FS_Config();
 //		импортирует свойства товаров из опций
-		if ( $_REQUEST['fs-api'] == 'migrate' ) {
+		if (isset( $_REQUEST['fs-api'] ) && $_REQUEST['fs-api'] == 'migrate' ) {
 			FS_Migrate_Class::import_option_attr();
 //			удаляет таблицу заказов
-		} elseif ( $_REQUEST['fs-api'] == 'drop_orders_table' ) {
+		} elseif (isset( $_REQUEST['fs-api'] ) && $_REQUEST['fs-api'] == 'drop_orders_table' ) {
 			$orders_table=$config->data['table_orders'];
 			$wpdb->query( "DROP TABLE IF EXISTS $orders_table") ;
